+++ ChannelIO/Source/VIPER/UserChat/UserChatInteractor.swift
@@ -108,7 +108,7 @@ class UserChatInteractor: NSObject, UserChatInteractorProtocol {
   
   func readyToPresent() -> Observable<Bool> {
     return Observable.create({ (subscriber) in
-      let signal = Observable.zip(CHPlugin.get(with: mainStore.state.plugin.id), CHManager.getRecentFollowers())
+      let signal = Observable.zip(CHPlugin.get(with: mainStore.state.plugin.id), AppManager.getOperators())
         .observeOn(MainScheduler.instance)
         .subscribe(onNext: { (info, managers) in
           mainStore.dispatch(UpdateFollowingManagers(payload: managers))
@@ -356,12 +356,6 @@ extension UserChatInteractor {
   }
   
   func send(message: CHMessage?) -> Observable<CHMessage?> {
-    return Observable.create({ (subscribe) -> Disposable in
-      return Disposables.create()
-    })
-  }
-  
-  func requestProfileBot() -> Observable<Bool?> {
     return Observable.create({ (subscribe) -> Disposable in
       return Disposables.create()
     })