+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 6.1.0
+## Updates
+* Changed some user chats' state
+* Hided delete message from chat
+* Added line integration
+* Updated InApp Push Notification view layout
+
 # 6.0.4
 ## Updates
 * Excluded UIAlertController from topController